backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
amqp==5.0.6
anyjson==0.3.3
arrow==1.2.1
asgiref==3.4.1
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
beautifulsoup4==4.10.0
bibtexparser==1.2.0
billiard==3.6.4.0
bs4==0.0.1
cached-property==1.5.2
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.9
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.9
django-celery==3.3.1
django-celery-beat==2.2.1
django-cors-headers==3.10.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
docutils==0.17.1
fake-useragent==0.1.11
free-proxy==1.0.4
frozenlist==1.2.0
future==0.18.2
idna==3.3
idna-ssl==1.1.0
imagesize==1.3.0
importlib-metadata==4.8.2
Jinja2==3.0.3
kombu==5.1.0
MarkupSafe==2.0.1
multidict==5.2.0
numpy==1.19.5
packaging==21.3
pandas==1.1.5
prompt-toolkit==3.0.23
pycodestyle==2.8.0
Pygments==2.10.0
PyJWT==1.7.1
pyparsing==3.0.6
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
scholarly==1.5.1
selenium==3.141.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==4.3.1
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
yarl==1.7.2
zipp==3.6.0
